setup: move cmdline functionality into common/cmdline.c file
Export parse_bool() as a static inline function.
