+++ include/setup.h
@@ -43,7 +43,6 @@ typedef struct boot_flags boot_flags_t;
 
 extern io_port_t com_ports[2];
 
-extern const char *kernel_cmdline;
 extern char cpu_identifier[49];
 extern boot_flags_t boot_flags;
